refactor(server): dedupe concurrency limiter, fix refresh-data error
