+++ server/src/utils/concurrency.ts
@@ -1,0 +1,15 @@
+export async function runWithConcurrencyLimit<T>(
+  tasks: Array<() => Promise<T>>,
+  concurrency: number
+): Promise<T[]> {
+  if (tasks.length === 0) {
+    return [];
+  }
+
+  const results: T[] = [];
+  for (let index = 0; index < tasks.length; index += concurrency) {
+    const chunk = tasks.slice(index, index + concurrency);
+    results.push(...(await Promise.all(chunk.map((task) => task()))));
+  }
+  return results;
+}
